requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.3
Django==5.1.6
django-cors-headers==4.7.0
django-crontab==0.7.1
djangorestframework==3.15.2
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
httplib2==0.22.0
idna==3.10
kombu==5.5.3
ndg-httpsclient==0.5.1
oauthlib==3.2.2
prompt_toolkit==3.0.51
psycopg2==2.9.10
pusher==3.3.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
